+++ src/parser/node_internal.rs
@@ -1,0 +1,7 @@
+use crate::tokenizer::Token;
+
+pub struct NodeInternal<'input> {
+    pub rule_name: String,
+    pub children: Vec<NodeInternal<'input>>,
+    pub token: Option<Token<'input>>,
+}
